Fixed shit so it works now.

+++ src/main/java/Backend/Rest/Repositories/CharacterRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends CrudRepository<Character, Long>{
+public interface CharacterRepository extends CrudRepository<Character, Integer>{
 
    List<Character> findByCharacterName(String name);
